Fixed a problem with units when using arbitrary plotting code

+++ lib/window.py
@@ -43,7 +43,9 @@ class Window(tk.Tk):
         
         # Remove all files from the "tmp" directory
         tmp_path = os.path.join(os.path.dirname(os.path.dirname(os.path.realpath(__file__))),"tmp")
-        for filename in os.listdir(tmp_path):
-            os.remove(os.path.join(tmp_path,filename))
+        if not os.path.isdir(tmp_path): os.mkdir(tmp_path)
+        else:
+            for filename in os.listdir(tmp_path):
+                os.remove(os.path.join(tmp_path,filename))
         
         close_window(self)
